Point index should only increment for valid points
This commit fixes a bug which would cause simplification to fail when
one or more points were dropped due to missing required points.

The problem was that index was incremented even when a point wasn't
pushed to `points`, which meant that a Point.index could be larger
than points.len().